Implement lazy loading in projects/new for unauthenticated users
ActionCable depends on authenticated users, so lazy-loading using a
job + turbo broadcast isn't feasible.

Instead, use a `turbo_frame_tag src:` to a route that responds with a
`turbo_stream.replace`.

This does lock up a web worker for a few seconds. For this project,
scalability (or even DoS attach) isn't a concern.