Delete Upsun configuration 😞
There are two issues not discussed in the documentation.

First, no ActiveJob runner is started. I'm not even sure that would be
possible without adding a separate worker, which would be overkill.

Second, ActionCable doesn't work because `GET /cable` can't be upgraded.
Maybe some tweaking of the web / location config could fix that?